+++ exonum-java-binding/cryptocurrency-demo/src/test/java/com/exonum/binding/cryptocurrency/CreateWalletTxIntegrationTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 The Exonum Team
+ * Copyright 2019 The Exonum Team
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,18 @@
  * limitations under the License.
  */
 
-package com.exonum.binding.cryptocurrency.transactions;
+package com.exonum.binding.cryptocurrency;
 
 import static com.exonum.binding.common.blockchain.ExecutionStatuses.serviceError;
-import static com.exonum.binding.cryptocurrency.transactions.PredefinedServiceParameters.ARTIFACT_FILENAME;
-import static com.exonum.binding.cryptocurrency.transactions.PredefinedServiceParameters.ARTIFACT_ID;
-import static com.exonum.binding.cryptocurrency.transactions.PredefinedServiceParameters.SERVICE_ID;
-import static com.exonum.binding.cryptocurrency.transactions.PredefinedServiceParameters.SERVICE_NAME;
-import static com.exonum.binding.cryptocurrency.transactions.PredefinedServiceParameters.artifactsDirectory;
-import static com.exonum.binding.cryptocurrency.transactions.TransactionError.WALLET_ALREADY_EXISTS;
-import static com.exonum.binding.cryptocurrency.transactions.TransactionUtils.DEFAULT_INITIAL_BALANCE;
-import static com.exonum.binding.cryptocurrency.transactions.TransactionUtils.newCreateWalletTransaction;
-import static com.exonum.binding.cryptocurrency.transactions.TransactionUtils.newCreateWalletTxPayload;
+import static com.exonum.binding.cryptocurrency.PredefinedServiceParameters.ARTIFACT_FILENAME;
+import static com.exonum.binding.cryptocurrency.PredefinedServiceParameters.ARTIFACT_ID;
+import static com.exonum.binding.cryptocurrency.PredefinedServiceParameters.SERVICE_ID;
+import static com.exonum.binding.cryptocurrency.PredefinedServiceParameters.SERVICE_NAME;
+import static com.exonum.binding.cryptocurrency.PredefinedServiceParameters.artifactsDirectory;
+import static com.exonum.binding.cryptocurrency.TransactionError.WALLET_ALREADY_EXISTS;
+import static com.exonum.binding.cryptocurrency.TransactionUtils.DEFAULT_INITIAL_BALANCE;
+import static com.exonum.binding.cryptocurrency.TransactionUtils.newCreateWalletTransaction;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import com.exonum.binding.common.crypto.KeyPair;
 import com.exonum.binding.common.crypto.PublicKey;
@@ -35,15 +33,11 @@ import com.exonum.binding.common.message.TransactionMessage;
 import com.exonum.binding.core.blockchain.Blockchain;
 import com.exonum.binding.core.storage.database.Snapshot;
 import com.exonum.binding.core.storage.indices.MapIndex;
-import com.exonum.binding.cryptocurrency.CryptocurrencySchema;
-import com.exonum.binding.cryptocurrency.PredefinedOwnerKeys;
-import com.exonum.binding.cryptocurrency.Wallet;
 import com.exonum.binding.test.RequiresNativeLibrary;
 import com.exonum.binding.testkit.TestKit;
 import com.exonum.binding.testkit.TestKitExtension;
 import com.exonum.core.messages.Runtime.ExecutionStatus;
 import java.util.Optional;
-import nl.jqno.equalsverifier.EqualsVerifier;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -58,26 +52,6 @@ class CreateWalletTxIntegrationTest {
           .withArtifactsDirectory(artifactsDirectory));
 
   private static final KeyPair OWNER_KEY_PAIR = PredefinedOwnerKeys.FIRST_OWNER_KEY_PAIR;
-
-  @Test
-  void from() {
-    long initialBalance = 100L;
-    byte[] arguments = newCreateWalletTxPayload(initialBalance);
-
-    CreateWalletTx tx = CreateWalletTx.from(arguments);
-
-    assertThat(tx).isEqualTo(new CreateWalletTx(initialBalance));
-  }
-
-  @Test
-  void constructorRejectsNegativeBalance() {
-    long initialBalance = -1L;
-
-    Throwable t = assertThrows(IllegalArgumentException.class,
-        () -> new CreateWalletTx(initialBalance));
-
-    assertThat(t.getMessage()).isEqualTo("The initial balance (-1) must not be negative.");
-  }
 
   @Test
   @RequiresNativeLibrary
@@ -119,12 +93,4 @@ class CreateWalletTxIntegrationTest {
     ExecutionStatus expectedTransactionResult = serviceError(WALLET_ALREADY_EXISTS.errorCode);
     assertThat(txResult).hasValue(expectedTransactionResult);
   }
-
-  @Test
-  void verifyEquals() {
-    EqualsVerifier
-        .forClass(CreateWalletTx.class)
-        .verify();
-  }
-
 }