+++ exonum-java-binding/fake-service/src/main/java/com/exonum/binding/fakeservice/FakeService.java
@@ -20,10 +20,14 @@ import com.exonum.binding.core.runtime.ServiceInstanceSpec;
 import com.exonum.binding.core.service.AbstractService;
 import com.exonum.binding.core.service.Node;
 import com.exonum.binding.core.storage.database.View;
+import com.exonum.binding.core.transaction.Transaction;
+import com.exonum.binding.core.transaction.TransactionContext;
 import com.google.inject.Inject;
 import io.vertx.ext.web.Router;
 
-final class FakeService extends AbstractService {
+public final class FakeService extends AbstractService {
+
+  static final int PUT_TX_ID = 0;
 
   @Inject
   FakeService(ServiceInstanceSpec instanceSpec) {
@@ -39,5 +43,18 @@ final class FakeService extends AbstractService {
   @Override
   public void createPublicApiHandlers(Node node, Router router) {
     // No handlers
+  }
+
+  /**
+   * Puts an entry (a key-value pair) into the test map.
+   */
+  @Transaction(PUT_TX_ID)
+  public void putEntry(Transactions.PutTransactionArgs arguments,
+      TransactionContext context) {
+    FakeSchema schema = new FakeSchema(context.getServiceName(), context.getFork());
+    String key = arguments.getKey();
+    String value = arguments.getValue();
+    schema.testMap()
+        .put(key, value);
   }
 }