+++ exonum-java-binding/qa-service/src/main/java/com/exonum/binding/qaservice/QaServiceModule.java
@@ -18,8 +18,6 @@ package com.exonum.binding.qaservice;
 
 import com.exonum.binding.core.service.AbstractServiceModule;
 import com.exonum.binding.core.service.Service;
-import com.exonum.binding.core.service.TransactionConverter;
-import com.exonum.binding.qaservice.transactions.QaTransactionConverter;
 import com.google.inject.Singleton;
 import org.pf4j.Extension;
 
@@ -35,7 +33,5 @@ public final class QaServiceModule extends AbstractServiceModule {
     bind(QaService.class).to(QaServiceImpl.class);
     // Make sure QaService remains a singleton.
     bind(QaServiceImpl.class).in(Singleton.class);
-
-    bind(TransactionConverter.class).to(QaTransactionConverter.class);
   }
 }